requirements-txt.txt
yt-dlp==2023.3.4
ffmpeg-python==0.2.0
openai-whisper==20230314
